--- Arrangement_on_surface_2/doc/Arrangement_on_surface_2/CGAL/IO/Qt_widget_Conic_arc_2.h
@@ -1,9 +1,0 @@
-/*!
-\ingroup PkgArrangement2op_left_shift
-Inserts a conic arc into a given `Qt_widget` stream.
-Only the basic geometric and topological features of the conic arcs are
-written. Auxiliary data that might be attached is lost.
-*/
-template<typename ConicArc>
-Qt_widget& operator<< (std::ostream& os,
-                       const _Conic_x_monotone_arc_2<ConicArc>& cv);
