+++ Surface_mesher/doc/Surface_mesher/CGAL/IO/Complex_2_in_triangulation_3_file_writer.h
@@ -29,5 +29,15 @@ bool output_surface_facets_to_off (std::ostream& os,
 				   int options = 
 				   Surface_mesher::IO_ORIENT_SURFACE);
 
+ namespace Surface_mesher {
+/*!
+\ingroup PkgSurfaceMesher3FunctionsIO
+ Options for function `output_surface_facets_to_off()`.
+*/
+    enum IO_option { NO_OPTION = 0,
+                     IO_ORIENT_SURFACE = 1,
+                     IO_VERBOSE = 2 };
+    
+  } /* namespace Surface_mesher */
 } /* namespace CGAL */
