--- Surface_mesher/doc/Surface_mesher/CGAL/IO/Io_option.h
@@ -1,13 +1,0 @@
-namespace CGAL {
-  namespace Surface_mesher {
- 
-/*!
-\ingroup PkgSurfaceMesher3FunctionsIO
- Options for function `output_surface_facets_to_off()`.
-*/
-    enum IO_option { NO_OPTION = 0,
-                     IO_ORIENT_SURFACE = 1,
-                     IO_VERBOSE = 2 };
-    
-  }
-}
